Move functions from utils to the Service
